+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is mostly based on [Keep a Changelog](https://keepachangelog.com/en/1
 ### Changed
 
 ### Fixed
+- Remove setting for minimum purge interval since it is not used.
 
 # Releases
 ## [18.3.0-beta1] - 2022-10-10